+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Navigation inside of group profile not linking to the sections page
+- Group profile title not changing from previous page
 
 ### Changed
 - Removed the content prop the `MiniCard` component